Provide a wrapper for c-function pointers
The intialization of encoding/decoding function fails when using
std::function<>. The reason for it is not fully clear and would
require more time to investigate.
As a work-around we can use a simple wrapper that provides only
the functionality that we require in a minimal way
to prevent access to null function pointers.